+++ crates/engine/tree/Cargo.toml
@@ -85,6 +85,7 @@ reth-static-file.workspace = true
 reth-testing-utils.workspace = true
 reth-tracing.workspace = true
 reth-trie-db.workspace = true
+reth-node-ethereum.workspace = true
 
 # alloy
 alloy-rlp.workspace = true
