+++ crates/ethereum/engine-primitives/Cargo.toml
@@ -12,11 +12,9 @@ workspace = true
 
 [dependencies]
 # reth
-reth-chainspec.workspace = true
 reth-primitives.workspace = true
 reth-engine-primitives.workspace = true
 reth-payload-primitives.workspace = true
-reth-payload-validator.workspace = true
 
 # alloy
 alloy-primitives.workspace = true
@@ -34,7 +32,6 @@ serde_json.workspace = true
 [features]
 default = ["std"]
 std = [
-    "reth-chainspec/std",
     "reth-primitives/std",
     "alloy-primitives/std",
     "alloy-eips/std",
