+++ crates/ethereum/node/Cargo.toml
@@ -33,9 +33,13 @@ reth-chainspec.workspace = true
 reth-revm = { workspace = true, features = ["std"] }
 reth-trie-db.workspace = true
 reth-rpc-eth-types.workspace = true
+reth-engine-primitives.workspace = true
+reth-primitives.workspace = true
+reth-payload-primitives.workspace = true
 
+# ethereum
 alloy-rpc-types-eth.workspace = true
-
+alloy-rpc-types-engine.workspace = true
 # revm with required ethereum features
 revm = { workspace = true, features = ["secp256k1", "blst", "c-kzg"] }
