+++ crates/rpc/rpc-builder/Cargo.toml
@@ -66,6 +66,7 @@ reth-transaction-pool = { workspace = true, features = ["test-utils"] }
 reth-rpc-types-compat.workspace = true
 reth-primitives.workspace = true
 reth-engine-primitives.workspace = true
+reth-node-ethereum.workspace = true
 
 alloy-primitives.workspace = true
 alloy-rpc-types-eth.workspace = true
