+++ crates/rpc/rpc-builder/tests/it/utils.rs
@@ -4,10 +4,11 @@ use alloy_rpc_types_engine::{ClientCode, ClientVersionV1};
 use reth_chainspec::MAINNET;
 use reth_consensus::noop::NoopConsensus;
 use reth_engine_primitives::BeaconConsensusEngineHandle;
-use reth_ethereum_engine_primitives::{EthEngineTypes, EthereumEngineValidator};
+use reth_ethereum_engine_primitives::EthEngineTypes;
 use reth_evm::execute::BasicBlockExecutorProvider;
 use reth_evm_ethereum::EthEvmConfig;
 use reth_network_api::noop::NoopNetwork;
+use reth_node_ethereum::EthereumEngineValidator;
 use reth_payload_builder::test_utils::spawn_test_payload_service;
 use reth_provider::test_utils::NoopProvider;
 use reth_rpc::EthApi;
